chore(workspace): more setup
added in contents from personal template, tweaked settings of nextjs template
