+++ .commitlintrc.js
@@ -1,0 +1,8 @@
+const { scopes } = require("./changelog.config");
+
+module.exports = {
+	"extends": ["@commitlint/config-conventional"],
+	"rules": {
+		"scope-enum": [2, "always", scopes],
+	},
+};
