+++ .eslintrc.json
@@ -28,6 +28,6 @@
 			}
 		],
 		"quotes": ["error", "single"],
-		"jsdoc/require-jsdoc": "off"
+		"jsdoc/require-jsdoc": 0
 	}
 }
