+++ .lintstagedrc.json
@@ -1,0 +1,5 @@
+{
+	"src/**/*.+(js|json|ts|tsx)": ["eslint --fix"],
+	"src/**/*.{css,scss,md}": ["prettier --write"],
+	"README.md": ["prettier --write"]
+}
