+++ changelog.config.js
@@ -1,0 +1,3 @@
+module.exports = {
+	scopes: ['workspace', 'backend', 'scripts', 'frontend', 'mixed'],
+};
