--- test/__mocks__/fileMock.js
@@ -1,1 +1,0 @@
-module.exports = 'test-file-stub'
